refactor: extract `contrapunk-harmony` crate
10 files (engine, scale, modes, key_detect, stateful, voice_leading/,
functional/, barry_harris, config). Internal cross-imports rewritten
`crate::harmony::*` → `crate::*`. Re-exported as `contrapunk::harmony`.
