+++ crates/contrapunk-harmony/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "contrapunk-harmony"
+version = "1.1.0"
+edition = "2021"
+description = "Harmony engine: keys, modes, voice leading, counterpoint, key detection"
+authors = ["Contrapunk Team"]
+
+[lib]
+name = "contrapunk_harmony"
+path = "src/lib.rs"
+
+[dependencies]
+wmidi = "4.0"
+rand = "0.10"
+serde = { version = "1.0", features = ["derive"] }
