+++ crates/contrapunk-harmony/src/lib.rs
